@@ -1,5 +1,5 @@
 module github.com/leon-domingo/get-latest-wp
 
-go 1.17
+go 1.20
 
 require github.com/buger/jsonparser v1.1.1
